+++ pi/display/layout.py
@@ -1,0 +1,3 @@
+
+class Layout:
+    hello_label = (10, 10)